ADD layer data export
Added a command line flag to export layers in a given directory. Not
that it's incredibly useful, but this allowed to build layer
decrypting/decoding in isolation. Which, given how painful it was, was a
good thing. And thus I learned that `char *` are not a proper
representation of byte arrays (had to use `u_int8_t *`, for some
reason).