CHANGE readme : remove mention of issue tracker
I disabled the issue tracker on all my github projects. It doesn't work:
it's not used as a forum for users to discuss their problems as I
wanted, it's still used by people expecting answers from me, probably
having not seen the maintenance notice in the readme, and probably
getting frustrated for not getting answers. It's better for everyone to
just deactive the issue tracker, especially since it still allows to
send pull requests.